1.21 port & 2.2.2.3 update

+++ src/main/java/io/github/lightman314/lightmanscurrency/client/gui/widget/button/trade/display/CoinPriceEntry.java
@@ -48,7 +48,7 @@ public class CoinPriceEntry extends DisplayEntry {
                 int left = this.getTopLeft(x + pos.xOffset(), pos.width());
                 int top = this.getTopLeft(y + pos.yOffset(), pos.height());
                 ItemStack stack = new ItemStack(entries.get(i).coin);
-                stack.setCount((int)entries.get(i).amount);
+                stack.setCount(entries.get(i).amount);
                 gui.renderItem(stack, left, top);
             }
         }
@@ -61,7 +61,7 @@ public class CoinPriceEntry extends DisplayEntry {
             for(int i = entries.size() - 1; i >= 0; --i)
             {
                 ItemStack stack = new ItemStack(entries.get(i).coin);
-                stack.setCount((int)entries.get(i).amount);
+                stack.setCount(entries.get(i).amount);
                 gui.renderItem(stack, left, top);
                 left -= spacing;
             }
